+++ backend/serverless.yml
@@ -11,6 +11,10 @@ provider:
   name: aws
   runtime: nodejs12.x
 
+  #tracing:
+   #   lambda: true
+    #  apiGateway: true
+
   stage: ${opt:stage, 'dev'}
   region: ${opt:region, 'us-east-2'}
